pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-langchain-backend"
version = "0.1.0"
description = "A FastAPI backend project with LangChain integration."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "langchain>=1.2.0,<2.0.0",
    "langchain-openai>=1.1.0,<2.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.fastapi]
entrypoint = "app.main:app"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
